Fix potential bug in options handling
Problem: If an option contained 0 or another false-ish value, it would
go ignored. So far this was no problem because no such values are to
be expected, but this might cause bugs if more options are introduced.

Solution: Check on the actual presence of the option key in the options
object regardless of value.

See discussion: https://twitter.com/felixge/status/55615426058649600